@@ -1,27 +1,32 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Amp\ParallelFunctions;
 
-use Amp\MultiReasonException;
-use Amp\Parallel\Worker\Pool;
-use Amp\Promise;
+use Amp\CompositeException;
+use Amp\Future;
+use Amp\Parallel\Worker\WorkerPool;
 use Amp\Serialization\SerializationException;
+use Closure;
 use Laravel\SerializableClosure\SerializableClosure;
-use function Amp\call;
-use function Amp\Parallel\Worker\enqueue;
-use function Amp\Promise\any;
+
+use function Amp\async;
+use function Amp\Future\awaitAll;
+use function Amp\Parallel\Worker\submit;
 
 /**
  * Parallelizes a callable.
  *
- * @param callable  $callable Callable to parallelize.
- * @param Pool|null $pool Worker pool instance to use or null to use the global pool.
+ * @template TReturn
  *
- * @return callable Callable executing in another thread / process.
+ * @param (callable(mixed...): TReturn)  $callable Callable to parallelize.
+ * @param WorkerPool|null $pool Worker pool instance to use or null to use the global pool.
+ *
+ * @return (Closure(mixed...): TReturn) Callable executing in another thread / process.
  * @throws SerializationException If the passed callable is not safely serializable.
  */
-function parallel(callable $callable, Pool $pool = null): callable
+function parallel(callable $callable, ?WorkerPool $pool = null): Closure
 {
+    /** @psalm-suppress DocblockTypeContradiction https://github.com/vimeo/psalm/issues/10029 */
     if ($callable instanceof \Closure) {
         $callable = new SerializableClosure($callable);
     }
@@ -32,81 +37,85 @@ function parallel(callable $callable, Pool $pool = null): callable
         throw new SerializationException("Unsupported callable: " . $e->getMessage(), 0, $e);
     }
 
-    return function (...$args) use ($pool, $callable): Promise {
+    return function (...$args) use ($pool, $callable): mixed {
         $task = new Internal\SerializedCallableTask($callable, $args);
-        return $pool ? $pool->enqueue($task) : enqueue($task);
+        return ($pool ? $pool->submit($task) : submit($task))->getFuture()->await();
     };
 }
 
 /**
  * Parallel version of array_map, but with an argument order consistent with the filter function.
  *
- * @param array     $array
- * @param callable  $callable
- * @param Pool|null $pool Worker pool instance to use or null to use the global pool.
+ * @template Tk as array-key
+ * @template TValue
+ * @template TReturn
  *
- * @return Promise Resolves to the result once the operation finished.
+ * @param array<Tk, TValue>     $array
+ * @param (callable(TValue): TReturn)|(callable(): TReturn)  $callable
+ * @param WorkerPool|null $pool Worker pool instance to use or null to use the global pool.
+ *
+ * @return array<Tk, TReturn> Resolves to the result once the operation finished.
  * @throws \Error If the passed callable is not safely serializable.
+ * @throws CompositeException If at least one call throws an exception.
  */
-function parallelMap(array $array, callable $callable, Pool $pool = null): Promise
+function parallelMap(array $array, callable $callable, ?WorkerPool $pool = null): array
 {
-    return call(function () use ($array, $callable, $pool) {
-        // Amp\Promise\any() guarantees that all operations finished prior to resolving. Amp\Promise\all() doesn't.
-        // Additionally, we return all errors as a MultiReasonException instead of throwing on the first error.
-        [$errors, $results] = yield any(\array_map(parallel($callable, $pool), $array));
+    /** @psalm-suppress PossiblyInvalidArgument Ignore this, we can't represent a variable number of args with templates */
+    $callable = parallel($callable, $pool);
+    $callable = fn (mixed $v): Future => async($callable, $v);
+    [$errors, $results] = awaitAll(\array_map($callable, $array));
 
-        if ($errors) {
-            throw new MultiReasonException($errors);
-        }
+    if ($errors) {
+        throw new CompositeException($errors);
+    }
 
-        return $results;
-    });
+    return $results;
 }
 
 /**
  * Parallel version of array_filter.
  *
- * @param array     $array
- * @param callable  $callable
- * @param int       $flag
- * @param Pool|null $pool Worker pool instance to use or null to use the global pool.
+ * @template Tk as array-key
+ * @template TValue
  *
- * @return Promise
+ * @param array<Tk, mixed> $array
+ * @param WorkerPool|null $pool Worker pool instance to use or null to use the global pool.
+ *
  * @throws \Error If the passed callable is not safely serializable.
+ * @throws CompositeException If at least one call throws an exception.
+ *
+ * @return array<Tk, TValue>
  */
-function parallelFilter(array $array, callable $callable = null, int $flag = 0, Pool $pool = null): Promise
+function parallelFilter(array $array, ?callable $callable = null, int $flag = 0, ?WorkerPool $pool = null): array
 {
-    return call(function () use ($array, $callable, $flag, $pool) {
-        if ($callable === null) {
-            if ($flag === \ARRAY_FILTER_USE_BOTH || $flag === \ARRAY_FILTER_USE_KEY) {
-                throw new \Error('A valid $callable must be provided if $flag is set.');
-            }
-
-            $callable = function ($value) {
-                return (bool) $value;
-            };
+    if ($callable === null) {
+        if ($flag === \ARRAY_FILTER_USE_BOTH || $flag === \ARRAY_FILTER_USE_KEY) {
+            throw new \Error('A valid $callable must be provided if $flag is set.');
         }
 
-        // Amp\Promise\any() guarantees that all operations finished prior to resolving. Amp\Promise\all() doesn't.
-        // Additionally, we return all errors as a MultiReasonException instead of throwing on the first error.
-        if ($flag === \ARRAY_FILTER_USE_BOTH) {
-            [$errors, $results] = yield any(\array_map(parallel($callable, $pool), $array, \array_keys($array)));
-        } elseif ($flag === \ARRAY_FILTER_USE_KEY) {
-            [$errors, $results] = yield any(\array_map(parallel($callable, $pool), \array_keys($array)));
-        } else {
-            [$errors, $results] = yield any(\array_map(parallel($callable, $pool), $array));
-        }
+        $callable = fn ($v): bool => (bool) $v;
+    }
+    $callable = parallel($callable, $pool);
+    $callable = fn (mixed ...$v): Future => async($callable, ...$v);
 
-        if ($errors) {
-            throw new MultiReasonException($errors);
-        }
+    if ($flag === \ARRAY_FILTER_USE_BOTH) {
+        /** @psalm-suppress TooManyArguments Ignore this, we can't represent a variable number of args with templates */
+        [$errors, $results] = awaitAll(\array_map($callable, $array, \array_keys($array)));
+    } elseif ($flag === \ARRAY_FILTER_USE_KEY) {
+        [$errors, $results] = awaitAll(\array_map($callable, \array_keys($array)));
+    } else {
+        [$errors, $results] = awaitAll(\array_map($callable, $array));
+    }
 
-        foreach ($array as $key => $arg) {
-            if (!$results[$key]) {
-                unset($array[$key]);
-            }
-        }
+    if ($errors) {
+        throw new CompositeException($errors);
+    }
 
-        return $array;
-    });
+    foreach ($array as $key => $_) {
+        if (!$results[$key]) {
+            unset($array[$key]);
+        }
+    }
+
+    return $array;
 }
