@@ -1,11 +1,10 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Amp\ParallelFunctions\Test;
 
-use Amp\MultiReasonException;
+use Amp\CompositeException;
 use Amp\PHPUnit\AsyncTestCase;
 use function Amp\ParallelFunctions\parallelFilter;
-use function Amp\Promise\wait;
 
 class FilterTest extends AsyncTestCase
 {
@@ -13,7 +12,7 @@ class FilterTest extends AsyncTestCase
     {
         $input = [1, 0, 3, false, true, null];
 
-        $this->assertSame(\array_filter($input), wait(parallelFilter($input)));
+        $this->assertSame(\array_filter($input), parallelFilter($input));
     }
 
     public function testWithCallback()
@@ -23,7 +22,7 @@ class FilterTest extends AsyncTestCase
             return $value === false;
         };
 
-        $this->assertSame(\array_filter($input, $callback), wait(parallelFilter($input, $callback)));
+        $this->assertSame(\array_filter($input, $callback), parallelFilter($input, $callback));
     }
 
     public function testWithCallbackAndFlagKey()
@@ -33,7 +32,7 @@ class FilterTest extends AsyncTestCase
             return $key === 2;
         };
 
-        $this->assertSame(\array_filter($input, $callback, \ARRAY_FILTER_USE_KEY), wait(parallelFilter($input, $callback, \ARRAY_FILTER_USE_KEY)));
+        $this->assertSame(\array_filter($input, $callback, \ARRAY_FILTER_USE_KEY), parallelFilter($input, $callback, \ARRAY_FILTER_USE_KEY));
     }
 
     public function testWithCallbackAndFlagBoth()
@@ -43,16 +42,16 @@ class FilterTest extends AsyncTestCase
             return $key === 2 || $value === true;
         };
 
-        $this->assertSame(\array_filter($input, $callback, \ARRAY_FILTER_USE_BOTH), wait(parallelFilter($input, $callback, \ARRAY_FILTER_USE_BOTH)));
+        $this->assertSame(\array_filter($input, $callback, \ARRAY_FILTER_USE_BOTH), parallelFilter($input, $callback, \ARRAY_FILTER_USE_BOTH));
     }
 
     public function testException()
     {
-        $this->expectException(MultiReasonException::class);
+        $this->expectException(CompositeException::class);
 
-        wait(parallelFilter([1, 2, 3], function () {
+        parallelFilter([1, 2, 3], function () {
             throw new \Exception;
-        }));
+        });
     }
 
     public function testExecutesAllTasksOnException()
@@ -64,7 +63,7 @@ class FilterTest extends AsyncTestCase
         ];
 
         try {
-            wait(parallelFilter($files, function ($args) {
+            parallelFilter($files, function ($args) {
                 list($id, $filename) = $args;
 
                 if ($id === 0) {
@@ -73,10 +72,10 @@ class FilterTest extends AsyncTestCase
 
                 \sleep(1);
                 \file_put_contents($filename, $id);
-            }));
+            });
 
             $this->fail('No exception thrown.');
-        } catch (MultiReasonException $e) {
+        } catch (CompositeException $e) {
             $this->assertStringEqualsFile($files[1][1], '1');
             $this->assertStringEqualsFile($files[2][1], '2');
         }
@@ -92,6 +91,6 @@ class FilterTest extends AsyncTestCase
             [2, \tempnam(\sys_get_temp_dir(), 'amp-parallel-functions-')],
         ];
 
-        wait(parallelFilter($files, null, ARRAY_FILTER_USE_BOTH));
+        parallelFilter($files, null, ARRAY_FILTER_USE_BOTH);
     }
 }
