@@ -1,37 +1,36 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Amp\ParallelFunctions\Test;
 
-use Amp\MultiReasonException;
+use Amp\CompositeException;
 use Amp\PHPUnit\AsyncTestCase;
 use function Amp\ParallelFunctions\parallelMap;
-use function Amp\Promise\wait;
 
 class MapTest extends AsyncTestCase
 {
     public function testValidInput()
     {
-        $this->assertSame([3, 4, 5], wait(parallelMap([1, 2, 3], function ($input) {
+        $this->assertSame([3, 4, 5], parallelMap([1, 2, 3], function ($input) {
             return $input + 2;
-        })));
+        }));
     }
 
     public function testCorrectOutputOrder()
     {
-        $this->assertSame([0, 1, 0], wait(parallelMap([0, 1, 0], function ($input) {
+        $this->assertSame([0, 1, 0], parallelMap([0, 1, 0], function ($input) {
             \sleep($input);
 
             return $input;
-        })));
+        }));
     }
 
     public function testException()
     {
-        $this->expectException(MultiReasonException::class);
+        $this->expectException(CompositeException::class);
 
-        wait(parallelMap([1, 2, 3], function () {
+        parallelMap([1, 2, 3], function () {
             throw new \Exception;
-        }));
+        });
     }
 
     public function testExecutesAllTasksOnException()
@@ -43,7 +42,7 @@ class MapTest extends AsyncTestCase
         ];
 
         try {
-            wait(parallelMap($files, function ($args) {
+            parallelMap($files, function ($args) {
                 list($id, $filename) = $args;
 
                 if ($id === 0) {
@@ -52,10 +51,10 @@ class MapTest extends AsyncTestCase
 
                 \sleep(1);
                 \file_put_contents($filename, $id);
-            }));
+            });
 
             $this->fail('No exception thrown.');
-        } catch (MultiReasonException $e) {
+        } catch (CompositeException $e) {
             $this->assertStringEqualsFile($files[1][1], '1');
             $this->assertStringEqualsFile($files[2][1], '2');
         }
